Change the way Monlog lib behaves
Act as python logging. You create a new logger instance. This instance
is connected to the caller, so every call to log() will be signed by the
callers class, i.e. "se.liu.tdp024.logic.bean.AccountBean".

Added newline safe way to log, that means newlines is converted to "\\n"
in the string.